refactor(mcp): move outbound session identity into auth
StartupIdentity, RequestScopedIdentity, and the AuthSource union resolve the outbound httpx.Auth a session acts as: StartupIdentity via pipefy_auth credential resolution, RequestScopedIdentity via auth's own require_request_bearer. They lived in core/runtime.py only because the runtime consumes them. Move them to auth/session_identity.py (the outbound counterpart to request_identity.py's inbound bearer) and publish them from auth's contract, so runtime.py is just the composition root: own the engine, wire resources per profile, open sessions. core imports the AuthSource contract inward from auth; the layering contract holds.
