+++ packages/mcp/src/pipefy_mcp/auth/__init__.py
@@ -1,4 +1,4 @@
-"""Resource-server auth for the HTTP transport: inbound bearer validation."""
+"""Identity for the HTTP transport: inbound bearer validation and session identity."""
 
 from pipefy_mcp.auth.request_identity import require_request_bearer
 from pipefy_mcp.auth.resource_server import (
@@ -7,11 +7,19 @@ from pipefy_mcp.auth.resource_server import (
     ResourceServerAuth,
     build_resource_server_auth,
 )
+from pipefy_mcp.auth.session_identity import (
+    AuthSource,
+    RequestScopedIdentity,
+    StartupIdentity,
+)
 
 __all__ = [
+    "AuthSource",
     "JwtTokenVerifier",
+    "RequestScopedIdentity",
     "ResourceServer",
     "ResourceServerAuth",
+    "StartupIdentity",
     "build_resource_server_auth",
     "require_request_bearer",
 ]
